[#36875831] assigned activities showing teacher pages and teacher notes.
https://www.pivotaltracker.com/story/show/36875831

Tests included.
One thing I don't understand: activities and pages controller seem to set
teacher mode by ORing params with .teacher_only? --

I am guessing that if the activity is marked as teacher_only it assumes
we must be in teacher mode.  Its a weird assumption.  I am going to keep
that in for now, but it something to notice.